requirements.txt
discord.py[voice]
python-dotenv
spotipy
sqlalchemy
yt-dlp
